Zoltan v3.9: From Trilinos hash c00ff3bb3393fb652581ec3695a6b5c401487be4 May 3, 2021

+++ src/driver/dr_gnuplot.c
@@ -45,6 +45,7 @@
  */
 
 #include "dr_const.h"
+#include "dr_externs.h"
 #include "dr_util_const.h"
 #include "dr_par_util_const.h"
 #include "dr_err_const.h"
@@ -127,14 +128,14 @@ int output_gnu(const char *cmd_file,
   int    i, j, nelems;
   int    prev_part = -1;
   int    max_part = -1;
-  float    locMaxX = INT_MIN;
-  float    locMinX = INT_MAX;
-  float    locMaxY = INT_MIN;
-  float    locMinY = INT_MAX;
-  float    globMaxX = INT_MIN;
-  float    globMinX = INT_MAX;
-  float    globMaxY = INT_MIN;
-  float    globMinY = INT_MAX;
+  float    locMaxX = (float)INT_MIN;
+  float    locMinX = (float)INT_MAX;
+  float    locMaxY = (float)INT_MIN;
+  float    locMinY = (float)INT_MAX;
+  float    globMaxX = (float)INT_MIN;
+  float    globMinX = (float)INT_MAX;
+  float    globMaxY = (float)INT_MIN;
+  float    globMinY = (float)INT_MAX;
   int    gmax_part = Num_Proc-1;
   int    gnum_part = Num_Proc;
   int   *parts = NULL;
